chore: split off arrays/owned
